first steps in new structure, definitions of physical processes outside user code, photon processes
