+++ src/ompmc.h
@@ -1,0 +1,357 @@
+/******************************************************************************
+ ompMC - An hybrid parallel implementation for Monte Carlo particle transport
+ simulations
+ 
+ Copyright (C) 2018 Edgardo Doerner (edoerner@fis.puc.cl)
+
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*****************************************************************************/
+
+/* Flag set by '--verbose' argument */
+extern int verbose_flag;
+
+/******************************************************************************/
+/* A simple C/C++ class to parse input files and return requested key value 
+https://github.com/bmaynard/iniReader */
+
+const int BUFFER_SIZE = 256;
+const int INPUT_PAIRS = 80;
+const char* INPUT_EXT = ".inp";  // extension of input files
+
+/* Parse a configuration file */
+extern void parseInputFile(char *file_name);
+
+/* Copy the value of the selected input item to the char pointer */
+extern int getInputValue(char *dest, char *key);
+
+/* Returns nonzero if line is a string containing only whitespace or is empty */
+extern int lineBlack(char *line);
+
+/* Remove white spaces from string str_untrimmed and saves the results in
+ str_trimmed. Useful for string input values, such as file names */
+extern void removeSpaces(char* str_trimmed, const char* str_untrimmed);
+
+struct inputItems {
+    char key[BUFFER_SIZE];
+    char value[BUFFER_SIZE];
+};
+
+extern struct inputItems input_items[];     // key,value pairs
+extern int input_idx;                       // number of key,value pair
+
+/******************************************************************************/
+
+/*******************************************************************************
+* Implementation, based on the EGSnrc one, of the RANMAR random number 
+* generator (RNG), proposed by Marsaglia and Zaman. 
+* 
+* Following the EGSnrc implementation, it uses integers to store the state of 
+* the RNG and to generate the next number in the sequence. Only at the end the 
+* random numbers are converted to floating point numbers, due to performance 
+* reasons. 
+* 
+* Before using the RNG, it is needed to initialize the RNG by a call to 
+* initRandom(). 
+*******************************************************************************/
+
+const int NRANDOM = 128;    // number of random numbers generated in each call 
+                            // to setRandom().
+
+struct Random {
+    int crndm;
+    int cdrndm;
+    int cmrndm;
+    int ixx;
+    int jxx;
+    int rng_seed;
+    int *urndm;
+    int *rng_array;
+    double twom24;
+};
+
+extern struct Random rng;
+#pragma omp threadprivate(rng)
+
+/* Initialization function for the RANMAR random number generator (RNG) 
+proposed by Marsaglia and Zaman and adapted from the EGSnrc version to be 
+used in ompMC. */
+extern void initRandom(void);
+
+/* Generation function for the RANMAR random number generator (RNG) proposed 
+by Marsaglia and Zaman. It generates NRANDOM floating point numbers in 
+each call */
+extern void getRandom(void);
+
+/* Get a single floating random number in [0,1) using the RANMAR RNG */
+extern double setRandom(void);
+
+extern void cleanRandom(void);
+
+/******************************************************************************/
+
+/*******************************************************************************
+/* Definitions for Monte Carlo simulation of particle transport 
+*******************************************************************************/
+
+/* Physical constants */
+const double RM = 0.5109989461; // MeV * c^(-2)
+
+/* Common functions and definitions */
+const int MXSTACK = 100;    // maximum number of particles in stack
+
+struct Stack {
+    int np;         // stack pointer
+    
+    int *iq;        // particle charge
+    int *ir;        // current region
+    
+    double *e;      // total particle energy
+    
+    double *x;      // particle coordinates
+    double *y;
+    double *z;
+    
+    double *u;      // particle direction cosines
+    double *v;
+    double *w;
+    
+    double *dnear;  // perpendicular distance to nearest boundary
+    double *wt;     // particle weight
+};
+extern struct Stack stack;
+#pragma omp threadprivate(stack)
+
+extern void initStack(void);
+extern void cleanStack(void);
+
+struct Uphi {
+    /* This structure holds data saved between uphi() calls */
+    double A, B, C;
+    double cosphi, sinphi;
+};
+
+struct Mscats {
+    /* This structure holds data saved between mscat calls */
+    int i;
+    int j;
+    double omega2;
+};
+
+struct Spinr {
+    /* This structure holds data saved between spinRejection calls */
+    int i;
+    int j;
+};
+
+extern void transferProperties(int npnew, int npold);
+extern void selectAzimuthalAngle(double *costhe, double *sinthe);
+extern void uphi21(struct Uphi *uphi, double costhe, double sinthe);
+extern void uphi32(struct Uphi *uphi, double costhe, double sinthe);
+
+/*******************************************************************************
+/* Photon physical processes definitions
+*******************************************************************************/
+
+const int MXGE = 2000;  // gamma mapped energy intervals
+
+struct Photon {
+    double *ge0, *ge1;
+    double *gmfp0, *gmfp1;
+    double *gbr10, *gbr11;
+    double *gbr20, *gbr21;
+    double *cohe0, *cohe1;
+};
+extern struct Photon photon_data;
+
+extern void readXsecData(char *file, int *ndat,
+                  double **xsec_data0,
+                  double **xsec_data1);
+
+extern void heap_sort(int n, double *values, int *indices);
+extern double *get_data(int flag, int ne, int *ndat,
+                 double **data0, double **data1,
+                 double *z_sorted, double *pz_sorted,
+                 double ge0, double ge1);
+extern double kn_sigma0(double e);
+
+extern void initPhotonData(void);
+extern void cleanPhoton(void);
+extern void listPhoton(void);
+
+/* Rayleigh scattering definitions */
+const int MXRAYFF = 100;        // Rayleigh atomic form factor
+const int RAYCDFSIZE = 100;     // CDF from Rayleigh form factors squared
+const double HC_INVERSE = 80.65506856998;
+const double TWICE_HC2 = 0.000307444456;
+
+struct Rayleigh {
+    double *xgrid;
+    double *fcum;
+    double *b_array;
+    double *c_array;
+    double *pmax0;
+    double *pmax1;
+    int *i_array;
+};
+extern struct Rayleigh rayleigh_data;
+
+extern void readFfData(double *xval, double **aff);
+extern void initRayleighData(void);
+extern void cleanRayleigh(void);
+extern void listRayleigh(void);
+
+/* Pair production definitions */
+const double FSC = 0.00729735255664;    // fine structure constant
+
+struct Pair {
+    double *dl1;
+    double *dl2;
+    double *dl3;
+    double *dl4;
+    double *dl5;
+    double *dl6;
+    
+    double *bpar0;
+    double *bpar1;
+    double *delcm;
+    double *zbrang;
+};
+
+extern struct Pair pair_data;
+
+extern double fcoulc(double zi);
+extern double xsif(double zi, double fc);
+extern void initPairData(void);
+extern void cleanPair(void);
+extern void listPair(void);
+
+/*******************************************************************************
+/* Electron physical processes definitions
+*******************************************************************************/
+
+struct Electron {
+    double *esig0;
+    double *esig1;
+    double *psig0;
+    double *psig1;
+    
+    double *ededx0;
+    double *ededx1;
+    double *pdedx0;
+    double *pdedx1;
+    
+    double *ebr10;
+    double *ebr11;
+    double *pbr10;
+    double *pbr11;
+    
+    double *pbr20;
+    double *pbr21;
+    
+    double *tmxs0;
+    double *tmxs1;
+    
+    double *blcce0;
+    double *blcce1;
+    
+    double *etae_ms0;
+    double *etae_ms1;
+    double *etap_ms0;
+    double *etap_ms1;
+    
+    double *q1ce_ms0;
+    double *q1ce_ms1;
+    double *q1cp_ms0;
+    double *q1cp_ms1;
+    
+    double *q2ce_ms0;
+    double *q2ce_ms1;
+    double *q2cp_ms0;
+    double *q2cp_ms1;
+    
+    double *range_ep;
+    double *e_array;
+    
+    double *eke0;
+    double *eke1;
+    
+    int *sig_ismonotone;
+    
+    double *esig_e;
+    double *psig_e;
+    double *xcc;
+    double *blcc;
+    double *expeke1;
+    
+};
+extern struct Electron electron_data;
+
+extern void shower(void);
+
+/* Media definitions */
+const int MXMED = 9;        // maximum number of media supported by the platform
+const int MXELEMENT = 50;   // maximum number of elements in a single medium
+const int MXEKE = 500;      // electron mapped energy intervals
+
+struct Media {
+    int nmed;                   // number of media in the problem
+    char med_names[MXMED][60];  // media names
+};
+extern struct Media media;
+
+struct Element {
+    /* Attributes of an element in a medium */
+    char symbol[3];
+    double z;
+    double wa;
+    double pz;
+    double rhoz;
+    
+};
+
+struct Pegs {
+    /* Data extracted from pegs file */
+    char names[MXMED][60];          // media names (as found in .pegs4dat file)
+    int ne[MXMED];                  // number of elements in medium
+    int iunrst[MXMED];              // flag for type of stopping power
+    int epstfl[MXMED];              // flag for ICRU37 collision stopping powers
+    int iaprim[MXMED];              // flag for ICRU37 radiative stopping powers
+    
+    int msge[MXMED];
+    int mge[MXMED];
+    int mseke[MXMED];
+    int meke[MXMED];
+    int mleke[MXMED];
+    int mcmfp[MXMED];
+    int mrange[MXMED];
+    
+    double rho[MXMED];              // mass density of medium
+    double rlc[MXMED];              // radiation length for the medium (in cm)
+    double ae[MXMED], ap[MXMED];    // electron and photon creation threshold E
+    double ue[MXMED], up[MXMED];    // upper electron and photon energy
+    double te[MXMED];
+    double thmoll[MXMED];
+    double delcm[MXMED];
+    
+    struct Element elements[MXMED][MXELEMENT];  // element properties
+};
+extern struct Pegs pegs_data;
+
+extern void initMediaData(void);
+extern int readPegsFile(int *media_found);
+
+
+/******************************************************************************/
+
